Move tests to scippneutron
