+++ tests/scippnexus/nxdetector_test.py
@@ -1,7 +1,7 @@
 import h5py
 import numpy as np
 import scipp as sc
-from scippnexus import NXroot, NX_class, NexusStructureError, File
+from scippnexus import NXroot, NX_class, NexusStructureError
 import pytest
 
 
@@ -189,36 +189,6 @@ def test_loading_event_data_with_full_selection_and_automatic_detector_numbers_w
     assert detector.dims == ['detector_number']
     assert detector[...].shape == [4]
     assert detector[()].shape == [4]
-
-
-@pytest.mark.skip("TODO")
-def test_can_load_nxdetector_from_bigfake():
-    import scippneutron as scn
-    with File(scn.data.bigfake()) as f:
-        da = f['entry/instrument/detector_1'][...]
-        assert da.sizes == {'dim_0': 300, 'dim_1': 300}
-
-
-@pytest.mark.skip("TODO")
-def test_can_load_nxdetector_from_PG3():
-    import scippneutron as scn
-    with File(scn.data.get_path('PG3_4844_event.nxs')) as f:
-        det = f['entry/instrument/bank24']
-        da = det[...]
-        assert da.sizes == {'x_pixel_offset': 154, 'y_pixel_offset': 7}
-        assert 'detector_number' not in da.coords
-        assert da.coords['pixel_id'].sizes == da.sizes
-        assert da.coords['distance'].sizes == da.sizes
-        assert da.coords['polar_angle'].sizes == da.sizes
-        assert da.coords['azimuthal_angle'].sizes == da.sizes
-        assert da.coords['x_pixel_offset'].sizes == {'x_pixel_offset': 154}
-        assert da.coords['y_pixel_offset'].sizes == {'y_pixel_offset': 7}
-        # local_name is an example of a dataset with shape=[1] that is treated as scalar
-        assert da.coords['local_name'].sizes == {}
-        # Extra scalar fields not in underlying NXevent_data
-        del da.coords['local_name']
-        del da.coords['total_counts']
-        assert sc.identical(da.sum(), det.events[()].sum())  # no event lost in binning
 
 
 def test_event_data_field_dims_labels(nxroot):
